first working cli rework
